Completed Makefile and fixed errors in functions. Added bonus functions to header

+++ srcs/ft_memcmp.c
@@ -6,7 +6,7 @@
 /*   By: bpuschel <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/11/03 19:01:26 by bpuschel          #+#    #+#             */
-/*   Updated: 2016/11/05 17:13:57 by bpuschel         ###   ########.fr       */
+/*   Updated: 2016/11/06 16:52:35 by bpuschel         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 
 int	ft_memcmp(const void *s1, const void *s2, size_t n)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	while (i < n)
